requirements.txt
advanced_alchemy==0.26.2
aiosqlite==0.20.0
alembic==1.14.0
anyio==4.7.0
certifi==2024.12.14
click==8.1.8
colorama==0.4.6
EditorConfig==0.17.0
Faker==33.1.0
fast-query-parsers==1.0.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jsbeautifier==1.15.1
litestar==2.13.0
litestar-htmx==0.4.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.18.6
multidict==6.1.0
polyfactory==2.18.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
rich==13.9.4
rich-click==1.8.5
ruff==0.8.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.3
websockets==14.1
